backend/requirements.txt
Django==5.0.6
djangorestframework==3.15.2
drf-spectacular==0.27.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.4.0
dj-database-url==2.3.0
psycopg2-binary==2.9.10
gunicorn==23.0.0
pytest==8.3.4
pytest-django==4.9.0
pytest-cov==6.0.0
